update view response

+++ Domain/Model/FarmActivity.cs
@@ -6,7 +6,7 @@ namespace Domain.Model;
 
 public partial class FarmActivity
 {
-    public int FarmActivitiesId { get; set; }
+    public long FarmActivitiesId { get; set; }
 
     public int? ActivityType { get; set; }
